Fixing client registration of multiblocks
When blocks updated their metadata, they trigger an invalidate/validate
thrash on the client (which they shouldn't, but do anyway because hey).
Working around this with some filthy logic in the desc-packet receipt,
which will always occur after an invalidate/validate thrash.

Also adding a world-unload callback so that we properly clean up
multiblock controllers in SSP